+++ runway/cfngin/context.py
@@ -42,14 +42,19 @@ class Context(object):
 
     """
 
-    def __init__(self, environment=None,
+    def __init__(self,
+                 environment=None,
+                 boto3_credentials=None,
                  stack_names=None,
                  config=None,
+                 config_path=None,
                  region=None,
                  force_stacks=None):
         """Instantiate class.
 
         Args:
+            boto3_credentials (Optional[Dict[str, str]]): Credentials to use
+                when creating a boto3 session from context.
             environment (dict): A dictionary used to pass in information about
                 the environment. Useful for templating.
             stack_names (list): A list of stack_names to operate on. If not
@@ -57,11 +62,13 @@ class Context(object):
                 operated on.
             config (:class:`runway.cfngin.config.Config`): The CFNgin
                 configuration being operated on.
+            config_path (str): Path to the config file that was provided.
             region (str): Name of an AWS region if provided as a CLI argument.
             force_stacks (list): A list of stacks to force work on. Used to
                 work on locked stacks.
 
         """
+        self.__boto3_credentials = boto3_credentials
         self._bucket_name = None
         self._persistent_graph = None
         self._persistent_graph_lock_code = None
@@ -70,12 +77,19 @@ class Context(object):
         self._stacks = None
         self._targets = None
         self._upload_to_s3 = None
+        # TODO load the config from context instead of taking it as an arg
         self.config = config or Config()
+        # TODO set this value when provisioning a Config object in context
+        # set to a fake location for the time being but this should be set
+        # by all runtime entry points. the only time the fake value should be
+        # used is during tests.
+        self.config_path = config_path or './'
         self.bucket_region = self.config.cfngin_bucket_region or region
         self.environment = environment
         self.force_stacks = force_stacks or []
         self.hook_data = {}
-        self.s3_conn = get_session(self.bucket_region).client('s3')
+        self.region = region
+        self.s3_conn = self.get_session(region=self.bucket_region).client('s3')
         self.stack_names = stack_names or []
 
     @property
@@ -302,6 +316,34 @@ class Context(object):
                 targets.append(target)
             self._targets = targets
         return self._targets
+
+    def get_session(self, profile=None, region=None):
+        """Create a thread-safe boto3 session.
+
+        Args:
+            profile (Optional[str]): The profile for the session.
+            region (Optional[str]): The region for the session.
+
+        Returns:
+            :class:`boto3.session.Session`: A thread-safe boto3 session.
+
+        """
+        kwargs = {}
+        if profile:
+            kwargs['profile'] = profile
+        elif self.__boto3_credentials:
+            kwargs.update({
+                'access_key': self.__boto3_credentials.get(
+                    'aws_access_key_id'
+                ),
+                'secret_key': self.__boto3_credentials.get(
+                    'aws_secret_access_key'
+                ),
+                'session_token': self.__boto3_credentials.get(
+                    'aws_session_token'
+                )
+            })
+        return get_session(region=region or self.region, **kwargs)
 
     def get_stack(self, name):
         """Get a stack by name.